pyproject.toml
[project]
name = "sentiment-analysis"
version = "0.1.0"
description = ""
authors = [
    {name = "Irina Pavlova"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "kagglehub (>=0.3.10,<0.4.0)",
    "fire (>=0.7.0,<0.8.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "pytorch-lightning (>=2.5.1.post0,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "numpy (==1.26.4)",
    "pyparsing (>=3.2.3,<4.0.0)",
    "cycler (>=0.12.1,<0.13.0)",
    "kiwisolver (==1.3.1)",
    "tensorboard (>=2.19.0,<3.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "omegaconf (>=2.3.0,<3.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "torch (>=2.7.0,<3.0.0)",
    "mlflow (>=2.22.0,<3.0.0)",
    "dvc (>=3.59.2,<4.0.0)",
    "onnxruntime (==1.15.1)",
    "onnx (>=1.18.0,<2.0.0)",
    "tensorrt (>=10.11.0.33,<11.0.0.0)",
    "seaborn (>=0.13.2,<0.14.0)"
]

[tool.poetry]
packages = [
    { include = "project" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
